LayoutTests/imported/w3c/web-platform-tests/file-system-access/sandboxed_FileSystemBaseHandle-postMessage-BroadcastChannel.https.window-expected.txt


FAIL Send and receive messages using a broadcast channel in an iframe, dedicated worker and service worker. assert_equals: The message target must receive a "broadcast-channel-created" message response. expected (string) "broadcast-channel-created" but got (undefined) undefined
